refactor: 인수테스트 대체
